backend/requirement.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
langchain
langchain-openai
langchain-community
google-generativeai
chromadb
python-dotenv